emitting events from composer for tasks instead of writing out to the commandline directly
