+++ package.json
@@ -24,9 +24,6 @@
     "test": "mocha"
   },
   "dependencies": {
-    "ansi-green": "^0.1.1",
-    "ansi-red": "^0.1.1",
-    "ansi-yellow": "^0.1.1",
     "bach": "^0.4.1",
     "chokidar": "^1.0.5",
     "component-emitter": "^1.2.0",
@@ -34,6 +31,10 @@
     "lazy-cache": "^0.1.0"
   },
   "devDependencies": {
+    "ansi-cyan": "^0.1.1",
+    "ansi-green": "^0.1.1",
+    "ansi-red": "^0.1.1",
+    "ansi-yellow": "^0.1.1",
     "bluebird": "^2.9.34",
     "engine-handlebars": "^0.6.1",
     "gulp-extname": "^0.2.0",
